Move Offer views to separate file
